norm

+++ src/ui/button.c
@@ -1,6 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   button.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tdelage <marvin@42.fr>                     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/01/28 06:15:05 by tdelage           #+#    #+#             */
+/*   Updated: 2025/01/28 06:36:11 by tdelage          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <libft.h>
 #include <mlx/mmlx.h>
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <ui/buttons.h>
 #include <ui/vline.h>
@@ -27,7 +40,7 @@ void	button_update(void *mlx, struct s_button *b)
 	if (b->is_clicked)
 		return ;
 	mlx_mouse_get_pos(mlx, &x, &y);
-	if (b->on_click && collide((int[4]){x, y, 0, 0}, (int *)b))
+	if (b->on_click && collide((int [4]){x, y, 0, 0}, (int *)b))
 	{
 		b->is_clicked = true;
 		if (b->has_idx)
@@ -43,13 +56,13 @@ void	button_draw(struct s_mlx *mlx, struct s_button *b)
 	int			y;
 	mlx_color	color;
 
-	color = (mlx_color) BORDERC;
+	color = (mlx_color)BORDERC;
 	mlx_mouse_get_pos(mlx->context, &x, &y);
-	if (collide((int[4]){x, y, 0, 0}, (int *)b))
-		color = (mlx_color) BORDER2C;
+	if (collide((int [4]){x, y, 0, 0}, (int *)b))
+		color = (mlx_color)(uint32_t)BORDER2C;
 	x = b->x + b->width / 2 - ((ft_strlen(b->text) * PPC) / 2);
 	y = b->y + b->height / 2 - (PPC / 2) + PPC - 2;
-	mlx_string_put(mlx->context, mlx->ui.win, x, y, (mlx_color) TEXTC, b->text);
+	mlx_string_put(mlx->context, mlx->ui.win, x, y, (mlx_color)TEXTC, b->text);
 	vline_print(mlx, (struct s_vline){b->x, b->y, b->y + b->height}, color);
 	vline_print(mlx, (struct s_vline){b->x + b->width, b->y, b->y + b->height},
 		color);
